NP-40217 Move project helper function
