+++ src/pages/public_registration/PublicFundingsContent.tsx
@@ -3,8 +3,7 @@ import OpenInNewIcon from '@mui/icons-material/OpenInNew';
 import { useTranslation } from 'react-i18next';
 import { Funding } from '../../types/registration.types';
 import { getLanguageString } from '../../utils/translation-helpers';
-import { getNfrProjectUrl } from '../registration/description_tab/projects_field/projectHelpers';
-import { fundingSourceIsNfr } from '../../utils/registration-helpers';
+import { fundingSourceIsNfr, getNfrProjectUrl } from '../registration/description_tab/projects_field/projectHelpers';
 import { CristinApiPath } from '../../api/apiPaths';
 import { FundingSources } from '../../types/project.types';
 import { useFetchResource } from '../../utils/hooks/useFetchResource';
