+++ src/pages/registration/description_tab/FundingsField.tsx
@@ -10,8 +10,7 @@ import { FundingSources } from '../../../types/project.types';
 import { emptyFunding, Funding, Registration } from '../../../types/registration.types';
 import { useFetchResource } from '../../../utils/hooks/useFetchResource';
 import { getLanguageString } from '../../../utils/translation-helpers';
-import { getNfrProjectUrl } from './projects_field/projectHelpers';
-import { fundingSourceIsNfr } from '../../../utils/registration-helpers';
+import { fundingSourceIsNfr, getNfrProjectUrl } from './projects_field/projectHelpers';
 import { DescriptionFieldNames, SpecificFundingFieldNames } from '../../../types/publicationFieldNames';
 import { NfrProjectSearch } from '../../../components/NfrProjectSearch';
 import { dataTestId } from '../../../utils/dataTestIds';
